feat: major refactor
The biggest improvement is the loop will shutdown
automatically if no effects exist at the start of
the animation frame.

~~~

The `FrameFn` function type was used as the callback
type of calls like `raf(...)` and `raf.onStart(...)`,
but it's now replaced by the new `Rafz.Effect` type.

This function takes a `frame` object as its only argument,
instead of the old `(dt, clock)` arguments.

The `frame` argument uses the new `Rafz.Frame` type.
This object is shared between frames, so you shouldn't
mutate it and you should clone it if you want to
keep it around for some reason.

It's worth noting that `frame.clock` will equal zero
for synchronous effects (denoted as such when the `raf.sync`
method wraps a scheduling call). This wasn't the case in
previous versions.

~~~

To set an error handler, you should now pass a function to
the `raf.catch` method. In past versions, you would set the
`raf.catch` property directly, which is no longer supported.

~~~

The addon modules no longer extend `raf` with new
methods. Instead, you can import the addons directly
from the `rafz` main module and use them in such a
way that can be tree-shaked and *doesn't* require an
empty import like…

   import "rafz/timeout"

…before the addon can be used. Instead, you would do
the following:

   import { rafTimeout } from "rafz"

~~~

I've added a `rafPromise` function that works like
a `Promise.resolve().then(...)` call, but your callback
won't be called until the next animation frame.

~~~

There's also a few optimizations added to the queue
processing logic!